Simplified navigation logic
Tried an approach where I check the prev page on the server, then send it back and use some logic on the server rendered pages to determine the correct menu. Worked on local but not on the server, since chrome and other browsers prevent sending the previous page (the referrer) back to the server.
So after several hours of frustration I went for a run on Sunday morning and it occured to me that simply adding new routes that render the same page would be a simpler solution. Less dodgy logic to content with, which means more performant code and something that's more easily maintainable. Plus I don't have the burden of maintainng three instances of the same page.
This could still cause duplicate content problems but this site is set not to be indexed so who cares. If it was I could use canonical URLs. Anyway it works now.